requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
faiss-cpu
pypdf
pandas
sqlalchemy
python-multipart
langgraph